+++ packages/0x.js/src/utils/order_state_utils.ts
@@ -18,6 +18,8 @@ import {constants} from '../utils/constants';
 import {OrderFilledCancelledLazyStore} from '../stores/order_filled_cancelled_lazy_store';
 import {BalanceAndProxyAllowanceLazyStore} from '../stores/balance_proxy_allowance_lazy_store';
 
+const ACCEPTABLE_RELATIVE_ROUNDING_ERROR = 0.0001;
+
 export class OrderStateUtils {
     private balanceAndProxyAllowanceLazyStore: BalanceAndProxyAllowanceLazyStore;
     private orderFilledCancelledLazyStore: OrderFilledCancelledLazyStore;
@@ -67,7 +69,7 @@ export class OrderStateUtils {
             zrxTokenAddress, signedOrder.maker,
         );
         const filledTakerTokenAmount = await this.orderFilledCancelledLazyStore.getFilledTakerAmountAsync(orderHash);
-        const canceledTakerTokenAmount = await this.orderFilledCancelledLazyStore.getCancelledTakerAmountAsync(
+        const cancelledTakerTokenAmount = await this.orderFilledCancelledLazyStore.getCancelledTakerAmountAsync(
             orderHash,
         );
         const unavailableTakerTokenAmount = await exchange.getUnavailableTakerAmountAsync(orderHash);
@@ -78,6 +80,9 @@ export class OrderStateUtils {
                                                                    .dividedToIntegerBy(totalTakerTokenAmount);
         const fillableMakerTokenAmount = BigNumber.min([makerProxyAllowance, makerBalance]);
         const remainingFillableMakerTokenAmount = BigNumber.min(fillableMakerTokenAmount, remainingMakerTokenAmount);
+        const remainingFillableTakerTokenAmount = remainingFillableMakerTokenAmount
+                                                  .times(totalTakerTokenAmount)
+                                                  .dividedToIntegerBy(totalMakerTokenAmount);
         // TODO: Handle edge case where maker token is ZRX with fee
         const orderRelevantState = {
             makerBalance,
@@ -85,13 +90,14 @@ export class OrderStateUtils {
             makerFeeBalance,
             makerFeeProxyAllowance,
             filledTakerTokenAmount,
-            canceledTakerTokenAmount,
+            cancelledTakerTokenAmount,
             remainingFillableMakerTokenAmount,
+            remainingFillableTakerTokenAmount,
         };
         return orderRelevantState;
     }
     private validateIfOrderIsValid(signedOrder: SignedOrder, orderRelevantState: OrderRelevantState): void {
-        const unavailableTakerTokenAmount = orderRelevantState.canceledTakerTokenAmount.add(
+        const unavailableTakerTokenAmount = orderRelevantState.cancelledTakerTokenAmount.add(
             orderRelevantState.filledTakerTokenAmount,
         );
         const availableTakerTokenAmount = signedOrder.takerTokenAmount.minus(unavailableTakerTokenAmount);
@@ -112,6 +118,13 @@ export class OrderStateUtils {
             if (orderRelevantState.makerFeeProxyAllowance.eq(0)) {
                 throw new Error(ExchangeContractErrs.InsufficientMakerFeeAllowance);
             }
+        }
+        const minFillableTakerTokenAmountWithinNoRoundingErrorRange = signedOrder.takerTokenAmount
+                                                                      .dividedBy(ACCEPTABLE_RELATIVE_ROUNDING_ERROR)
+                                                                      .dividedBy(signedOrder.makerTokenAmount);
+        if (orderRelevantState.remainingFillableTakerTokenAmount
+            .lessThan(minFillableTakerTokenAmountWithinNoRoundingErrorRange)) {
+            throw new Error(ExchangeContractErrs.OrderFillRoundingError);
         }
         // TODO Add linear function solver when maker token is ZRX #badass
         // Return the max amount that's fillable