+++ packages/connect/src/index.ts
@@ -1,0 +1,11 @@
+export {HttpClient} from './http_client';
+export {
+    Client,
+    FeesRequest,
+    FeesResponse,
+    OrderbookRequest,
+    OrderbookResponse,
+    OrdersRequest,
+    TokenPairsItem,
+    TokenPairsRequest,
+} from './types';