+++ packages/connect/src/schemas/relayer_orders_request_schema.ts
@@ -1,0 +1,16 @@
+export const relayerOrdersRequestSchema = {
+    id: '/RelayerOrdersRequest',
+    type: 'object',
+    properties: {
+        exchangeContractAddress: {$ref: '/Address'},
+        tokenAddress: {$ref: '/Address'},
+        makerTokenAddress: {$ref: '/Address'},
+        takerTokenAddress: {$ref: '/Address'},
+        tokenA: {$ref: '/Address'},
+        tokenB: {$ref: '/Address'},
+        maker: {$ref: '/Address'},
+        taker: {$ref: '/Address'},
+        trader: {$ref: '/Address'},
+        feeRecipient: {$ref: '/Address'},
+    },
+};