+++ packages/connect/src/schemas/schemas.ts
@@ -1,0 +1,15 @@
+import {
+    relayerOrderBookRequestSchema,
+} from './relayer_orderbook_request_schema';
+import {
+    relayerOrdersRequestSchema,
+} from './relayer_orders_request_schema';
+import {
+    relayerTokenPairsRequestSchema,
+} from './relayer_token_pairs_request_schema';
+
+export const schemas = {
+    relayerOrderBookRequestSchema,
+    relayerOrdersRequestSchema,
+    relayerTokenPairsRequestSchema,
+};