+++ packages/connect/test/fixtures/standard_relayer_api/fees.ts
@@ -1,0 +1,8 @@
+import {BigNumber} from 'bignumber.js';
+import {FeesResponse} from '../../../src/types';
+
+export const feesResponse: FeesResponse = {
+    feeRecipient: '0x323b5d4c32345ced77393b3530b1eed0f346429d',
+    makerFee: new BigNumber('10000000000000000'),
+    takerFee: new BigNumber('30000000000000000'),
+};