+++ packages/connect/test/fixtures/standard_relayer_api/snapshot_orderbook_channel_message.ts
@@ -1,0 +1,17 @@
+import * as orderbookJSON from './orderbook.json';
+
+const orderbookJsonString = JSON.stringify(orderbookJSON);
+
+export const snapshotOrderbookChannelMessage = `{
+    "type": "snapshot",
+    "channel": "orderbook",
+    "channelId": 1,
+    "payload": ${orderbookJsonString}
+}`;
+
+export const malformedSnapshotOrderbookChannelMessage = `{
+    "type": "snapshot",
+    "channel": "orderbook",
+    "channelId": 1,
+    "payload": {}
+}`;