+++ packages/json-schemas/schemas/order_cancel_schema.ts
@@ -1,0 +1,12 @@
+export const orderCancellationRequestsSchema = {
+    id: '/OrderCancellationRequests',
+    type: 'array',
+    items: {
+        properties: {
+            order: {$ref: '/Order'},
+            takerTokenCancelAmount: {$ref: '/Number'},
+        },
+        required: ['order', 'takerTokenCancelAmount'],
+        type: 'object',
+    },
+};