+++ packages/json-schemas/schemas/signed_orders_schema.ts
@@ -1,0 +1,5 @@
+export const signedOrdersSchema = {
+    id: '/signedOrdersSchema',
+    type: 'array',
+    items: {$ref: '/SignedOrder'},
+};